DDCNL-1725 | ClosePostalAddressController created
