+++ app/views/cards/personaldetails/postalAddress.scala.html
@@ -53,7 +53,7 @@
                 </li>
                 @if(closePostalAddressEnabled) {
                     <li>
-                        <a class="ga-track-anchor-click" href="@controllers.routes.AddressController.closePostalAddressChoice" data-ga-event-category="link - click" data-ga-event-action="Postal address" data-ga-event-label="Send letters to your main address">@messages("label.send_your_letters_to_your_main_address")</a>
+                        <a class="ga-track-anchor-click" href="@controllers.address.routes.ClosePostalAddressController.onPageLoad()" data-ga-event-category="link - click" data-ga-event-action="Postal address" data-ga-event-label="Send letters to your main address">@messages("label.send_your_letters_to_your_main_address")</a>
                     </li>
                 }
             </ul>
