+++ test/controllers/AddressControllerSpec.scala
@@ -157,104 +157,6 @@ class AddressControllerSpec extends BaseSpec with MockitoSugar {
     })
   }
 
-  "Calling AddressController.closePostalAddressChoice" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(true)))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-
-    }
-
-    "display the closeCorrespondenceAddressChoice form that contains the main address" in new LocalSetup {
-      val result = controller.closePostalAddressChoice(FakeRequest())
-
-      contentAsString(result) should include(fakeAddress.line1.getOrElse("line6"))
-
-      status(result) shouldBe OK
-    }
-  }
-
-  "Calling AddressController.processClosePostalAddressChoice" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(true)))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-    }
-
-    "return 400 when supplied invalid form input" in new LocalSetup {
-
-      val result = controller.processClosePostalAddressChoice(FakeRequest("POST", ""))
-
-      status(result) shouldBe BAD_REQUEST
-    }
-
-    "return 303, caching closePostalAddressChoiceDto and redirecting to review changes page when supplied valid form input" in new LocalSetup {
-
-      val requestWithForm = FakeRequest("POST", "").withFormUrlEncodedBody("closePostalAddressChoice" -> "true")
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = requestWithForm.asInstanceOf[Request[A]])
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processClosePostalAddressChoice()(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/close-correspondence-address-confirm")
-    }
-
-    "return 303, caching closePostalAddressChoiceDto and redirecting to personal details page" in new LocalSetup {
-
-      val requestWithForm = FakeRequest("POST", "").withFormUrlEncodedBody("closePostalAddressChoice" -> "false")
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = requestWithForm.asInstanceOf[Request[A]])
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processClosePostalAddressChoice(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-    }
-  }
-
-  "Calling AddressController.confirmClosePostalAddress" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressLookupServiceDown" -> Json.toJson(Some(true)))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-
-    }
-
-    "render the appropriate content that includes the address" in new LocalSetup {
-      val result = controller.confirmClosePostalAddress(FakeRequest())
-
-      contentAsString(result) should include(fakeAddress.line1.getOrElse("line6"))
-    }
-  }
-
   "Calling AddressController.reviewChanges" should {
 
     trait LocalSetup extends WithAddressControllerSpecSetup {
@@ -422,266 +324,6 @@ class AddressControllerSpec extends BaseSpec with MockitoSugar {
 
       contentAsString(result) should include(Messages("label.your_address"))
       contentAsString(result) shouldNot include(Messages("label.when_you_started_living_here"))
-    }
-  }
-
-  "Calling AddressController.closePostalAddressChoice" should {
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-      override lazy val sessionCacheResponse = None
-
-    }
-
-    "return OK when closePostalAddressChoice is called" in new LocalSetup {
-
-      val result = controller.closePostalAddressChoice(FakeRequest())
-
-      status(result) shouldBe OK
-    }
-  }
-
-  "Calling AddressController.processClosePostalAddressChoice" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val sessionCacheResponse =
-        Some(CacheMap("id", Map("addressPageVisitedDto" -> Json.toJson(AddressPageVisitedDto(true)))))
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-    }
-
-    "redirect to expected confirm close correspondence confirmation page when supplied with value = Yes (true)" in new LocalSetup {
-      val requestWithForm: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", "")
-          .withFormUrlEncodedBody("closePostalAddressChoice" -> "true")
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = requestWithForm
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processClosePostalAddressChoice(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/your-address/close-correspondence-address-confirm")
-    }
-
-    "redirect to personal details page when supplied with value = No (false)" in new LocalSetup {
-      val requestWithForm: FakeRequest[AnyContentAsFormUrlEncoded] =
-        FakeRequest("POST", "")
-          .withFormUrlEncodedBody("closePostalAddressChoice" -> "false")
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              request = requestWithForm
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.processClosePostalAddressChoice(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some("/personal-account/personal-details")
-    }
-
-    "return a bad request when supplied no value" in new LocalSetup {
-      val result = controller.processClosePostalAddressChoice(FakeRequest())
-
-      status(result) shouldBe BAD_REQUEST
-    }
-  }
-
-  "Calling AddressController.confirmClosePostalAddress" should {
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddress
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(fakePersonDetails)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-      override lazy val sessionCacheResponse = None
-
-    }
-
-    "return OK when confirmClosePostalAddress is called" in new LocalSetup {
-      val result = controller.confirmClosePostalAddress(FakeRequest())
-
-      status(result) shouldBe OK
-    }
-  }
-
-  "Calling AddressController.submitConfirmClosePostalAddress" should {
-
-    trait LocalSetup extends WithAddressControllerSpecSetup {
-      override lazy val fakeAddress = buildFakeAddressWithEndDate
-      override lazy val nino = Fixtures.fakeNino
-      override lazy val personDetailsResponse = PersonDetailsSuccessResponse(buildPersonDetailsCorrespondenceAddress)
-      override lazy val updateAddressResponse: UpdateAddressResponse = UpdateAddressSuccessResponse
-      override lazy val thisYearStr = "2015"
-      override lazy val sessionCacheResponse = Some(CacheMap("id", Map.empty))
-
-      def comparatorDataEvent(dataEvent: DataEvent, auditType: String, uprn: Option[String]) = DataEvent(
-        "pertax-frontend",
-        auditType,
-        dataEvent.eventId,
-        Map("path" -> "/test", "transactionName" -> "closure_of_correspondence"),
-        Map(
-          "nino"              -> Some(Fixtures.fakeNino.nino),
-          "etag"              -> Some("115"),
-          "submittedLine1"    -> Some("1 Fake Street"),
-          "submittedLine2"    -> Some("Fake Town"),
-          "submittedLine3"    -> Some("Fake City"),
-          "submittedLine4"    -> Some("Fake Region"),
-          "submittedPostcode" -> Some("AA1 1AA"),
-          "submittedCountry"  -> None,
-          "addressType"       -> Some("correspondence")
-        ).collect { case (k, Some(v)) => k -> v },
-        dataEvent.generatedAt
-      )
-    }
-
-    "render the thank you page upon successful submission of closing the correspondence address" in new LocalSetup {
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              personDetails = Some(buildPersonDetailsCorrespondenceAddress),
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress(FakeRequest())
-
-      status(result) shouldBe OK
-
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(dataEvent, "closedAddressSubmitted", Some("GB101"))
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(1)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
-    }
-
-    "redirect to personal details if there is a lock on the correspondence address for the user" in new LocalSetup {
-      override def getEditedAddressIndicators: List[AddressJourneyTTLModel] =
-        List(mock[AddressJourneyTTLModel])
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = request, personDetails = Some(buildPersonDetailsCorrespondenceAddress))
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress(FakeRequest())
-
-      status(result) shouldBe SEE_OTHER
-      redirectLocation(result) shouldBe Some(routes.AddressController.personalDetails().url)
-
-      verify(mockAuditConnector, times(0)).sendEvent(any())(any(), any())
-      verify(mockCitizenDetailsService, times(0)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(0)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
-    }
-
-    "return 400 if UpdateAddressBadRequestResponse is received from citizen-details" in new LocalSetup {
-      override lazy val updateAddressResponse = UpdateAddressBadRequestResponse
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = request, personDetails = Some(buildPersonDetailsCorrespondenceAddress))
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress()(FakeRequest())
-
-      status(result) shouldBe BAD_REQUEST
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(0)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
-    }
-
-    "return 500 if an UpdateAddressUnexpectedResponse is received from citizen-details" in new LocalSetup {
-      override lazy val updateAddressResponse = UpdateAddressUnexpectedResponse(HttpResponse(SEE_OTHER))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = request, personDetails = Some(buildPersonDetailsCorrespondenceAddress))
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress()(FakeRequest())
-
-      status(result) shouldBe INTERNAL_SERVER_ERROR
-      verify(mockCitizenDetailsService, times(1))
-        .updateAddress(meq(Fixtures.fakeNino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(0)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
-    }
-
-    "return 500 if an UpdateAddressErrorResponse is received from citizen-details" in new LocalSetup {
-      override lazy val updateAddressResponse = UpdateAddressErrorResponse(new RuntimeException("Any exception"))
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(request = request, personDetails = Some(buildPersonDetailsCorrespondenceAddress))
-              .asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress()(FakeRequest())
-
-      status(result) shouldBe INTERNAL_SERVER_ERROR
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(0)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
-    }
-
-    "return 500 if insert address lock fails" in new LocalSetup {
-      override def isInsertCorrespondenceAddressLockSuccessful: Boolean = false
-
-      when(mockAuthJourney.authWithPersonalDetails).thenReturn(new ActionBuilderFixture {
-        override def invokeBlock[A](request: Request[A], block: UserRequest[A] => Future[Result]): Future[Result] =
-          block(
-            buildUserRequest(
-              saUser = NonFilerSelfAssessmentUser,
-              personDetails = Some(buildPersonDetailsCorrespondenceAddress),
-              request = FakeRequest("POST", "/test")
-                .asInstanceOf[Request[A]]
-            ).asInstanceOf[UserRequest[A]]
-          )
-      })
-
-      val result = controller.submitConfirmClosePostalAddress(FakeRequest())
-
-      status(result) shouldBe INTERNAL_SERVER_ERROR
-
-      val arg = ArgumentCaptor.forClass(classOf[DataEvent])
-      verify(mockAuditConnector, times(1)).sendEvent(arg.capture())(any(), any())
-      val dataEvent = arg.getValue
-
-      pruneDataEvent(dataEvent) shouldBe comparatorDataEvent(dataEvent, "closedAddressSubmitted", Some("GB101"))
-      verify(mockCitizenDetailsService, times(1)).updateAddress(meq(nino), meq("115"), meq(fakeAddress))(any())
-      verify(controller.editAddressLockRepository, times(1)).insert(meq(nino.withoutSuffix), meq(PostalAddrType))
     }
   }
